+++ apps/api-server/src/routes/api/idea.js
@@ -7,7 +7,7 @@ const db = require('../../db');
 const auth = require('../../middleware/sequelize-authorization-middleware');
 const mail = require('../../lib/mail');
 const pagination = require('../../middleware/pagination');
-const searchResults = require('../../middleware/search-results-static');
+const searchInResults = require('../../middleware/search-in-results');
 const isJson = require('../../util/isJson');
 const publishConcept = require('../../middleware/publish-concept');
 const c = require('config');
@@ -53,10 +53,6 @@ router
 
     if (req.query.mapMarkers) {
       req.scope.push('mapMarkers');
-    }
-
-    if (req.query.filters || req.query.exclude) {
-      req.scope.push({ method: ['filter', req.query.filters, req.query.exclude] });
     }
 
     if (req.query.running) {
@@ -105,7 +101,6 @@ router.route('/')
   // ----------
   .get(auth.can('Idea', 'list'))
   .get(pagination.init)
-  // add filters
   .get(function(req, res, next) {
     let { dbQuery } = req;
 
@@ -145,7 +140,7 @@ router.route('/')
       .catch(next);
   })
   .get(auth.useReqUser)
-  .get(searchResults)
+  .get(searchInResults({}))
   .get(pagination.paginateResults)
   .get(function(req, res, next) {
     res.json(req.results);