+++ apps/api-server/src/routes/api/user.js
@@ -6,7 +6,7 @@ const db = require('../../db');
 const auth = require('../../middleware/sequelize-authorization-middleware');
 const pagination = require('../../middleware/pagination');
 const {Op} = require('sequelize');
-const searchResults = require('../../middleware/search-results-user');
+const searchInResults = require('../../middleware/search-in-results');
 const fetch = require('node-fetch');
 const merge = require('merge');
 const authSettings = require('../../util/auth-settings');
@@ -47,39 +47,17 @@ router.route('/')
   })
   .get(pagination.init)
   .get(function (req, res, next) {
-    let {dbQuery} = req;
 
-    if (!dbQuery.where) {
-      dbQuery.where = {};
-    }
-
-    if (dbQuery.where.q) {
-      dbQuery.search = {
-        haystack: ['role', 'name'],
-        needle: dbQuery.where.q,
-        offset: dbQuery.offset,
-        limit: dbQuery.limit,
-        pageSize: dbQuery.pageSize,
-      };
-
-      delete dbQuery.where.q;
-      delete dbQuery.offset;
-      delete dbQuery.limit;
-      delete dbQuery.pageSize;
-    }
-
-    /**
-     * Add projectId to query conditions
-     * @type {{projectId: *}}
-     */
-    const queryConditions = Object.assign(dbQuery.where, {projectId: req.params.projectId});
+    let { dbQuery } = req;
+    dbQuery.where = {
+      projectId: req.params.projectId,
+      ...req.queryConditions,
+      ...dbQuery.where,
+    };
 
     db.User
       .scope(...req.scope)
-      .findAndCountAll({
-        ...dbQuery,
-        where: queryConditions,
-      })
+      .findAndCountAll(dbQuery)
       .then(function (result) {
         req.results = result.rows;
         req.dbQuery.count = result.count;
@@ -88,7 +66,7 @@ router.route('/')
       .catch(next);
   })
   .get(auth.useReqUser)
-  .get(searchResults)
+  .get(searchInResults({ searchfields: ['name', 'role'] }))
   .get(pagination.paginateResults)
   .get(function (req, res, next) {
     res.json(req.results);